[proj-49] feat: LoanSlider 컴포넌트 분리
